Visualization_service/pyproject.toml
[project]
name = "visualization-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Timon",email = "timonbub@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.3.3,<3.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "visualization_service", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
